+++ ios/PrayerApp/PrayerWidget.m
@@ -50,4 +50,19 @@ RCT_EXPORT_METHOD(setUiHints:(NSString *)style
   resolve(nil);
 }
 
+RCT_EXPORT_METHOD(setWidgetHighlightDynamic:(BOOL)enabled
+                  resolver:(RCTPromiseResolveBlock)resolve
+                  rejecter:(RCTPromiseRejectBlock)reject)
+{
+  NSUserDefaults *group =
+      [[NSUserDefaults alloc] initWithSuiteName:@"group.com.prayerapp"];
+  NSUserDefaults *target = group != nil ? group : [NSUserDefaults standardUserDefaults];
+  [target setBool:enabled forKey:@"widget_highlight_dynamic"];
+  [target synchronize];
+  if (@available(iOS 14.0, *)) {
+    [[WidgetCenter sharedWidgetCenter] reloadAllTimelines];
+  }
+  resolve(nil);
+}
+
 @end